+++ http4k-contract/src/main/kotlin/org/http4k/contract/openapi/v3/FieldRetrieval.kt
@@ -19,10 +19,10 @@ interface FieldRetrieval : (Any, String) -> Field {
     }
 }
 
-object SimpleLookup : FieldRetrieval {
+class SimpleLookup(private val renamingStrategy: (String) -> String = { it }) : FieldRetrieval {
     override fun invoke(target: Any, name: String): Field {
         val fields = try {
-            target::class.memberProperties.map { it.name to it }.toMap()
+            target::class.memberProperties.map { renamingStrategy(it.name) to it }.toMap()
         } catch (e: Error) {
             emptyMap<String, KProperty1<out Any, Any?>>()
         }
@@ -34,7 +34,6 @@ object SimpleLookup : FieldRetrieval {
                     ?.let { it to field.returnType.isMarkedNullable }
             }
             ?.let { Field(it.first, it.second) } ?: throw NoFieldFound(name, target)
-
     }
 }