+++ http4k-contract/src/test/kotlin/org/http4k/contract/openapi/v3/AutoJsonToJsonSchemaTest.kt
@@ -61,7 +61,7 @@ class AutoJsonToJsonSchemaTest {
     private val json = Jackson
 
     private val creator = AutoJsonToJsonSchema(json,
-        FieldRetrieval.compose(SimpleLookup, JacksonAnnotated),
+        FieldRetrieval.compose(SimpleLookup(), JacksonJsonPropertyAnnotated, JacksonJsonNamingAnnotated(Jackson)),
         SchemaModelNamer.Full,
         "customPrefix"
     )