+++ http4k-contract/src/test/kotlin/org/http4k/contract/openapi/v3/SimpleLookupTest.kt
@@ -11,13 +11,13 @@ class SimpleLookupTest {
 
     @Test
     fun `finds value from object`() {
-        assertThat("nonNullable", SimpleLookup(Beany(), "nonNullable"), equalTo(Field("hello", false)))
-        assertThat("aNullable", SimpleLookup(Beany(), "aNullable"), equalTo(Field("aNullable", true)))
+        assertThat("nonNullable", SimpleLookup()(Beany(), "nonNullable"), equalTo(Field("hello", false)))
+        assertThat("aNullable", SimpleLookup()(Beany(), "aNullable"), equalTo(Field("aNullable", true)))
     }
 
     @Test
     fun `throws on no field found`() {
-        assertThat("non existent", { SimpleLookup(Beany(), "non existent") }, throws<NoFieldFound>())
+        assertThat("non existent", { SimpleLookup()(Beany(), "non existent") }, throws<NoFieldFound>())
     }
 }